Refactor: input mode of desktop
